# cmd/arti-server/openapi.yaml
openapi: 3.0.3
info:
  title: arti
  version: 0.1.0
  description: |
    Artifact service — REST + MCP. Single Postgres table, S3-backed blobs,
    OIDC/session auth. Guides and reference render in-app at /help.
servers:
  - url: https://arti.example.com
  - url: http://localhost:8090
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
security:
  - bearerAuth: []
paths:
...
  /api/artifacts/{id}/files:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get: { summary: PACKAGE manifest, responses: { "200": { description: ok } } }
...